[graphite] Allow undefined for view format in getRenderPassDescKey
Per discussion on https://skia-review.googlesource.com/c/skia/+/874796
this also moves the write swizzle back out of the render pass desc key.
To preserve the reduced overall pipeline key size, this reduces the
allowed bits for texture formats by 1 so the RenderPassDesc key can
fit in 32-bits.

Bug: chromium:351964372
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/875796
Commit-Queue: Michael Ludwig <michaelludwig@google.com>

+++ src/gpu/graphite/dawn/DawnCaps.h
@@ -60,7 +60,7 @@ public:
                             ResourceType,
                             Shareable,
                             GraphiteResourceKey*) const override;
-    uint64_t getRenderPassDescKeyForPipeline(const RenderPassDesc& renderPassDesc) const;
+    uint32_t getRenderPassDescKeyForPipeline(const RenderPassDesc& renderPassDesc) const;
 
 private:
     const ColorTypeInfo* getColorTypeInfo(SkColorType, const TextureInfo&) const override;
